+++ examples/example-next-13-next-auth/src/components/LocaleSwitcher.tsx
@@ -1,13 +1,15 @@
-import Link from 'next/link';
 import {useLocale, useTranslations} from 'next-intl';
+import {usePathname} from 'next-intl/client';
+import Link from 'next-intl/link';
 
 export default function LocaleSwitcher() {
   const t = useTranslations('LocaleSwitcher');
   const locale = useLocale();
   const otherLocale = locale === 'en' ? 'de' : 'en';
+  const pathname = usePathname();
 
   return (
-    <Link href={'/' + otherLocale} prefetch={false}>
+    <Link href={pathname} locale={otherLocale}>
       {t('switchLocale', {locale: otherLocale})}
     </Link>
   );