+++ packages/next-intl/test/middleware/middleware.test.tsx
@@ -1,3 +1,5 @@
+// @vitest-environment edge-runtime
+
 import {RequestCookies} from 'next/dist/compiled/@edge-runtime/cookies';
 import {NextRequest, NextResponse} from 'next/server';
 import {it, describe, vi, beforeEach, expect, Mock} from 'vitest';
@@ -6,7 +8,7 @@ import {DomainConfig} from '../../src/middleware/NextIntlMiddlewareConfig';
 import {COOKIE_LOCALE_NAME} from '../../src/shared/constants';
 
 vi.mock('next/server', () => {
-  type MiddlewareResponseInit = Parameters<(typeof NextResponse)['next']>[0];
+  type MiddlewareResponseInit = Parameters<typeof NextResponse['next']>[0];
 
   function createResponse(init: MiddlewareResponseInit) {
     const response = new Response(null, init);
@@ -60,9 +62,9 @@ function createMockRequest(
 }
 
 const MockedNextResponse = NextResponse as unknown as {
-  next: Mock<Parameters<(typeof NextResponse)['next']>>;
-  rewrite: Mock<Parameters<(typeof NextResponse)['rewrite']>>;
-  redirect: Mock<Parameters<(typeof NextResponse)['redirect']>>;
+  next: Mock<Parameters<typeof NextResponse['next']>>;
+  rewrite: Mock<Parameters<typeof NextResponse['rewrite']>>;
+  redirect: Mock<Parameters<typeof NextResponse['redirect']>>;
 };
 
 beforeEach(() => {